requirements.txt
discord.py
beautifulsoup4
urllib3
requests
Pillow
